Move SdlTransport back to Android only
