--- baseAndroid/src/main/java/com/smartdevicelink/transport/SdlTransport.java
@@ -1,1 +1,0 @@
-../../../../../../../base/src/main/java/com/smartdevicelink/transport/SdlTransport.java
